fix build after commit 2c4e6ed3585ec63adee2e9a680de3c52b4db3c0c:
move SDL_UsingGameInputForXInputControllers() to core/windows/SDL_gameinput.cpp
because joystick/gdk/SDL_gameinputjoystick.cpp isn't always added to build.
