+++ src/joystick/windows/SDL_xinputjoystick.c
@@ -27,17 +27,16 @@
 #include "SDL_windowsjoystick_c.h"
 #include "SDL_xinputjoystick_c.h"
 #include "SDL_rawinputjoystick_c.h"
+#include "../../core/windows/SDL_gameinput.h"
 #include "../hidapi/SDL_hidapijoystick_c.h"
-#include "../gdk/SDL_gameinputjoystick_c.h"
 
 // Set up for C function definitions, even when using C++
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-/*
- * Internal stuff.
- */
+// Internal stuff
+
 static bool s_bXInputEnabled = false;
 
 bool SDL_XINPUT_Enabled(void)
